reatualizando as funcitons
