--- functions/favorite_activity.js
@@ -1,39 +1,0 @@
-const functions = require('firebase-functions');
-const admin = require('firebase-admin');
-admin.initializeApp();
-
-exports.onCreateFavoriteActivity = functions.firestore
-    .document('/posts_favoritos/{userId}/postsFavoritosUsuario/{documentId}')
-    .onCreate(async (snapshot, context) => {
-        const userId = context.params.userId;
-        const documentId = context.params.documentId;
-
-        //atualizar o conteudo na timeline
-        const timelineActivitiesRef = admin
-            .firestore()
-            .collection('timeline')
-            .doc(userId)
-            .collection('atividadesTimeline')
-            .doc(documentId);
-
-        timelineActivitiesRef.update({ "favorito": true });
-
-    });
-
-exports.onDeleteFavoriteActivity = functions.firestore
-    .document('/posts_favoritos/{userId}/postsFavoritosUsuario/{documentId}')
-    .onDelete(async (snapshot, context) => {
-        const userId = context.params.userId;
-        const documentId = context.params.documentId;
-
-        //atualizar o conteudo na timeline
-        const timelineActivitiesRef = admin
-            .firestore()
-            .collection('timeline')
-            .doc(userId)
-            .collection('atividadesTimeline')
-            .doc(documentId);
-
-        timelineActivitiesRef.update({ "favorito": false });
-
-    });
